Case 9000886: Changed the nav fonts to 12px.
Fonts too big for full menu, reduced to 12px (from 13px). For secondary nav drop down, reduced from 14px to 13px. Gave secondary navbar (navbar-nav) a left margin to move to the right of the logo. Gave the secondary nav a white background so text on page is not seen when scrolling. Add box shadow & border back in on drop down (until instructed differently). Removed standard 40px indent for unordered lists (as pushes log in and buy now buttons off the line for secondary nav). Gave the log in and buy now buttons a standard padding, so the size adjusts to text in button.