Case 9001003: Moved 'our specialties' to below 'why mexico insurance'
The design Chris provided did not go with the way we are doing feature icons on the rest of the site (we use blue with orange hover colors), he used organge icons. Rather than change everything on the site, I chose to move the feature icons below. CHanged the h2 to blue. Changed css for list elements on feature icons. Also changed the background color, color, opacity on the 'why mexico insurance' section. Made the text larger on the why mexico insurance section.